requirements.txt
# Core Whisper
openai-whisper>=20231117

# Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Audio Processing (via whisper dependencies)
# ffmpeg-python
# torch
# torchaudio
